useCase for report, refactor
