+++ domain/src/main/java/com/example/domain/model/HabitFulfillmentReport.kt
@@ -1,0 +1,3 @@
+package com.example.domain.model
+
+data class HabitFulfillmentReport(val type: Type, val difference: Int)
